requirements.txt
django~=3.1.2
pytz~=2020.1
sqlparse~=0.4.1
asgiref~=3.2.10
Pillow~=8.0.1
djangorestframework~=3.12.1
importlib-metadata~=2.0.0
markdown~=3.3.3
zipp~=3.4.0
django-filter~=2.4.0
django-rest-auth~=0.9.5
six~=1.15.0
certifi~=2020.6.20
cffi~=1.14.3
chardet~=3.0.4
cryptography~=3.2
defusedxml~=0.6.0
django-allauth~=0.43.0
idna~=2.10
oauthlib~=3.1.0
pycparser~=2.20
pyjwt~=1.7.1
python3-openid~=3.2.0
requests~=2.24.0
requests-oauthlib~=1.3.0
urllib3~=1.25.11



